backend/requirements.txt
livekit-agents[voice-assistant]
groq
livekit-plugins-silero
livekit-plugins-deepgram
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
gunicorn
